requirements.txt
boto3~=1.16.46
pyyaml~=5.3.1
aws-cdk.core~=1.86.0
aws-cdk.aws-certificatemanager~=1.86.0
aws-cdk.aws-cloudfront~=1.86.0
aws-cdk.aws-s3~=1.86.0
aws-cdk.aws-s3-notifications~=1.86.0
aws-cdk.aws-lambda~=1.86.0
aws-cdk.aws-iam~=1.86.0
aws-cdk.aws-cloudfront-origins~=1.86.0
aws-cdk.aws-route53~=1.86.0
aws-cdk.aws-route53-targets~=1.86.0
Click~=7.1.2
tqdm~=4.56.0